+++ Craftsman/Builders/Projects/WebApiCsProjBuilder.cs
@@ -77,7 +77,7 @@
     <PackageReference Include=""Serilog.AspNetCore"" Version=""4.1.0"" />
     <PackageReference Include=""Serilog.Enrichers.AspNetCore"" Version=""1.0.0"" />
     <PackageReference Include=""Serilog.Enrichers.Context"" Version=""4.2.0"" />
-    <PackageReference Include=""Serilog.Enrichers.Environment"" Version=""2.1.3"" />
+     <PackageReference Include=""Serilog.Exceptions"" Version=""7.1.0"" />
     <PackageReference Include=""Serilog.Enrichers.Process"" Version=""2.0.1"" />
     <PackageReference Include=""Serilog.Enrichers.Thread"" Version=""3.1.0"" />
     <PackageReference Include=""Serilog.Settings.Configuration"" Version=""3.1.0"" />